Removed indirections for all methods: moved implementations from DatabaseHandler to Sensor and Source.
