+++ sense-android-library/src/nl/sense_os/datastorageengine/RealmDatabaseHandler.java
@@ -11,7 +11,6 @@ import java.util.List;
 
 import io.realm.Realm;
 import io.realm.RealmResults;
-import nl.sense_os.service.shared.SensorDataPoint;
 
 
 /**
@@ -64,148 +63,11 @@ public class RealmDatabaseHandler implements DatabaseHandler {
     }
 
     /**
-     * This method inserts a DataPoint object into the local Realm database.
-     * The typecasting to string should already be done at this point.
-     * @param dataPoint	A DataPoint object that has a stringified value that will be copied
-     * 			into a Realm object.
-     */
-    protected void insertDataPoint (DataPoint dataPoint) {
-        RealmModelDataPoint realmDataPoint = RealmModelDataPoint.fromDataPoint(dataPoint);
-
-        realm.beginTransaction();
-        realm.copyToRealmOrUpdate(realmDataPoint);
-        realm.commitTransaction();
-    }
-
-    /**
-     * Get data points from the local database with the given sensor id.
-     * @param sensorId: String for the sensorID of the sensor that the data point belongs to.
-     * @param startDate: Start date of the query, included.
-     * @param endDate: End date of the query, excluded.
-     * @param limit: The maximum number of data points.
-     * @param sortOrder: Sort order, either ASC or DESC
-     * @return Returns a List with data points
-     */
-    protected List<DataPoint> getDataPoints(String sensorId, long startDate, long endDate, int limit, SORT_ORDER sortOrder) throws JSONException {
-        // query results
-        realm.beginTransaction();
-        RealmResults<RealmModelDataPoint> results = realm
-                .where(RealmModelDataPoint.class)
-                .equalTo("sensorId", sensorId)
-                .greaterThanOrEqualTo("date", startDate)
-                .lessThan("date", endDate)
-                .findAll();
-        realm.commitTransaction();
-
-        // sort
-        boolean resultsOrder = (sortOrder == SORT_ORDER.DESC )
-                ? RealmResults.SORT_ORDER_DESCENDING
-                : RealmResults.SORT_ORDER_ASCENDING;
-        results.sort("date", resultsOrder);
-
-        // limit and convert to DataPoint
-        int count = 0;
-        List<DataPoint> dataPoints = new ArrayList<>();
-        Iterator<RealmModelDataPoint> iterator = results.iterator();
-        while (count < limit && iterator.hasNext()) {
-            dataPoints.add(RealmModelDataPoint.toDataPoint(iterator.next()));
-            count++;
-        }
-        // TODO: figure out what is the most efficient way to loop over the results
-
-        return dataPoints;
-    }
-
-    /**
-     * Create a new Sensor and store it in the local database
-     */
-    protected Sensor createSensor(String id, String name, String userId, String sourceId, SensorDataPoint.DataType dataType, String csId, SensorOptions options, boolean synced) {
-        Sensor sensor = new RealmSensor(this, id, name, userId, sourceId, dataType, csId, options, synced);
-
-        RealmModelSensor realmSensor = RealmModelSensor.fromSensor(sensor);
-
-        realm.beginTransaction();
-        realm.copyToRealm(realmSensor);
-        realm.commitTransaction();
-
-        return sensor;
-    }
-
-    /**
-     * Update RealmSensor in local database with the info of the given Sensor object. Throws an exception if it fails to updated.
-     * @param sensor: Sensor object containing the updated info.
-     */
-    protected void updateSensor(Sensor sensor) throws DatabaseHandlerException {
-        realm.beginTransaction();
-
-        // get the existing sensor (just to throw an error if it doesn't exist)
-        RealmModelSensor realmSensor = realm
-                .where(RealmModelSensor.class)
-                .equalTo("id", sensor.getId())
-                .findFirst();
-        
-        if (realmSensor == null) {
-            throw new DatabaseHandlerException("Cannot update sensor: sensor doesn't yet exist.");
-        }
-
-        // update the sensor
-        realm.copyToRealmOrUpdate(RealmModelSensor.fromSensor(sensor));
-
-        realm.commitTransaction();
-    }
-
-    /**
-     * Get a sensor
-     * @param sourceId	The ID of the source object or Null
-     * @param sensorName	The name of the sensor or Null
-     * @return sensor: sensor with the given sensor name and sourceId.
-     **/
-    protected Sensor getSensor(String sourceId, String sensorName) throws JSONException {
-        realm.beginTransaction();
-
-        RealmModelSensor realmSensor = realm
-                .where(RealmModelSensor.class)
-                .equalTo("sourceId", sourceId)
-                .equalTo("name", sensorName)
-                .findFirst();
-
-        realm.commitTransaction();
-
-        return RealmModelSensor.toSensor(this, realmSensor);
-    }
-
-    /**
-     * Retrieve all sensors for given source id.
-     * @param sourceId
-     * @return
-     */
-    protected List<Sensor> getSensors(String sourceId) throws JSONException {
-        // query results
-        realm.beginTransaction();
-        RealmResults<RealmModelSensor> results = realm
-                .where(RealmModelSensor.class)
-                .equalTo("sourceId", sourceId)
-                .findAll();
-        realm.commitTransaction();
-
-        // convert to Sensor
-        List<Sensor> sensors = new ArrayList<>();
-        Iterator<RealmModelSensor> iterator = results.iterator();
-        while (iterator.hasNext()) {
-            sensors.add(RealmModelSensor.toSensor(this, iterator.next()));
-        }
-        // TODO: figure out what is the most efficient way to loop over the results
-
-        return sensors;
-    }
-
-
-    /**
      * Create a new source and store it in the local database
      */
     public Source createSource(String id, String name, JSONObject meta, String deviceId, String userId, String csId) {
         boolean synced = false;
-        Source source = new RealmSource(this, id, name, meta, deviceId, userId, csId, synced);
+        Source source = new RealmSource(realm, id, name, meta, deviceId, userId, csId, synced);
 
         RealmModelSource realmSource = RealmModelSource.fromSource(source);
 
@@ -214,30 +76,6 @@ public class RealmDatabaseHandler implements DatabaseHandler {
         realm.commitTransaction();
 
         return source;
-    }
-
-    /**
-     * Update Source in database with the info of the given Source object.
-     * Throws an exception if it fails to updated.
-     * @param source: Source object containing the updated info.
-     */
-    protected void updateSource(Source source) throws DatabaseHandlerException {
-        realm.beginTransaction();
-
-        // get the existing source (just to throw an error if it doesn't exist)
-        RealmModelSource realmSource = realm
-                .where(RealmModelSource.class)
-                .equalTo("id", source.getId())
-                .findFirst();
-
-        if (realmSource == null) {
-            throw new DatabaseHandlerException("Cannot update source: source doesn't yet exist.");
-        }
-
-        // update the source
-        realm.copyToRealmOrUpdate(RealmModelSource.fromSource(source));
-
-        realm.commitTransaction();
     }
 
     /**
@@ -260,11 +98,10 @@ public class RealmDatabaseHandler implements DatabaseHandler {
         List<Source> sources = new ArrayList<>();
         Iterator<RealmModelSource> iterator = results.iterator();
         while (iterator.hasNext()) {
-            sources.add(RealmModelSource.toSource(this, iterator.next()));
+            sources.add(RealmModelSource.toSource(realm, iterator.next()));
         }
         // TODO: figure out what is the most efficient way to loop over the results
 
         return sources;
     }
-
 }
